// oxygen-cms-ui/src/app/client/private/models/sport.model.ts
import { SvgFilename } from './svgfilename.model';
import { Base } from './base.model';

export interface Sport extends Base {
  alt: string;
  categoryId: number;
  collectionType: string;
  disabled: boolean;
  dispSortName: string;
  heightMedium: number;
  heightMediumIcon: number;
  heightSmall: number;
  heightSmallIcon: number;
  imageTitle: string;
  inApp: boolean;
  isOutrightSport: boolean;
  lang: string;
  outcomesTemplateType1: string;
  outcomesTemplateType2: string;
  outcomesTemplateType3: string;
  primaryMarkets: string;
  showInPlay: boolean;
  sortOrder: number;
  spriteClass: string;
  ssCategoryCode: string;
  svg: string;
  svgId: string;
  tabCompetitions: any;
  tabCoupons: any;
  tabJackpot: any;
  tabLive: any;
  tabMatches: any;
  tabOutrights: any;
  tabSpecials: any;
  targetUri: string;
  typeIds: string;
  uriMedium: string;
  uriSmall: string;
  viewByFilters: string;
  widthMedium: number;
  widthMediumIcon: number;
  widthSmall: number;
  widthSmallIcon: number;
  heightLarge: number;
  heightLargeIcon: number;
  uriLarge: string;
  defaultTab: string;
  widthLarge: number;
  widthLargeIcon: number;
  icon: SvgFilename;
  filename: SvgFilename;
  svgFilename: SvgFilename;
  fileNameString: string;
  isMultiTemplateSport?: boolean;
}